Coalesce BasicConstructClauses with shared variables.
If one or more BasicConstructClauses share variables, then they are
coalesced into a single BasicConstructClause in which the
ConstructPattern is the set union of all ConstructPatterns in the joined
set, while the WHEN condition is the conjunction of all conditions in the
joined set. Dupicate patterns and conditions are removed.