Find shortest paths with GraphX.
A Pregel-like algorithm is used to find shortest paths between all the
vertices in the graph and a subset of them, called the landmark
vertices. The graph data is converted from our internal representation
into the GraphX representation, then the SP algorithm is fired and,
finally, the result is converted back into our internal representation.