Fix CONSTRUCT and make it simpler.
- Remove the normalization of conditional constructs, as it was
semantically wrong to merge them.
- Replace the previous solution with a more lightweight one, in which we
only join the unmatched grouped vertices to the binding table. For
unmatched ungrouped vertices we still add a new column, but for the
matched vertices we no longer join back to the base table. Also, the new
identities are now added in the AddColumn operator, to simplify the
process. Finally, aggregate expressions are solved directly as
properties they were meant to be, no longer using an intermediary property in
the GroupBy.